Test(https://github.com/Tag-Me-DAW2/sprints/issues/34): Cambiado una parte de logica del test

+++ src/main/java/com/tagme/tagme_store_back/persistence/dao/jpa/impl/UserJpaDaoImpl.java
@@ -59,6 +59,10 @@ public class UserJpaDaoImpl implements UserJpaDao {
 
     @Override
     public void deleteById(Long id) {
+        UserJpaEntity managed = entityManager.find(UserJpaEntity.class, id);
+        if (managed == null) {
+            throw new ResourceNotFoundException("User not found");
+        }
         entityManager.remove(entityManager.find(UserJpaEntity.class, id));
     }
